Fixes Keras layer reparameterization in eager mode.
Storing a "computed" layer parameter like a reparameterized kernel as an object
attribute in the `build()` function breaks backpropagation in eager mode. The
reason is that if `build()` is called outside (before) a `GradientTape` scope,
the gradients from the layer output to its variables will be disconnected.

One fix is to make `Parameterizer`s return an object that, like variables, has
a `value()` method, and define the layer's parameters via `@property`s. Every
time a value of a parameter is requested, these then call the `value()` method
and thus recompute the parameter from its variable every time the layer is
called. This way, the dependency can be tracked with the current `GradientTape`
scope.

PiperOrigin-RevId: 292501734
Change-Id: Idec76b516f7799c32e5b0375b09eb76a63e0372f