Implements distribution objects for entropy modeling.
This reimplements the distributions from our existing entropy models in
`entropy_models.py` as separate classes. They tie in to the
`tensorflow_probability` class hierarchy.

The `DeepFactorized` class implements the density model currently implemented in
`EntropyBottleneck`. The `*Conditional` classes, which are based on existing
standard distributions and modified by adding uniform noise, are now implemented
via a `UniformNoiseAdapter` distribution, which takes a base distribution as
argument and implements the additive noise modification. It has some similarity
to the `QuantizedDistribution` class in `tensorflow_probability`, and some parts
of it are based on code I took from there. Making the additive noise
modification modular should enable experimentation with a larger variety of
density models more easily.

PiperOrigin-RevId: 293462740
Change-Id: I275d2c7a3bb8177424546755492be774b5222bf9