Implements entropy models for continuous-valued random variables.
This reimplements the existing entropy models in `entropy_models.py` in a more
modular way. The `EntropyBottleneck` class is now implemented through
`ContinuousBatchedEntropyModel` and can be instantiated with arbitrary
`Distribution` objects (use `DeepFactorized` to get the equivalent behavior to
`EntropyBottleneck`).

The old `*Conditional` classes are implemented via
`ContinuousIndexedEntropyModel` and the special case
`LocationScaleIndexedEntropyModel`. Like the batched version, they can be
instantiated with arbitrary distribution objects (use `NoisyNormal` to get the
equivalent behavior to `GaussianConditional`, etc.). In addition,
`ContinuousIndexedEntropyModel` now also supports multi-dimensional indexing.
The use case for that is to model conditionally independent scalar distributions
which have more than one parameter that depends on data (e.g., a shape and scale
parameter).

PiperOrigin-RevId: 293468780
Change-Id: I8189ccd898d8ae33819c36feff7fbc971b8523d0