Makes entropy models stateless.
The stateful implementation of the entropy models was broken. For instance, if
the model was set up like this:

```
log_scale = tf.Variable()
scale = tf.exp(log_scale)
em = ContinuousBatchedEntropyModel(Normal(scale=scale))
```

A subsequent training loop would not be able to backpropagate into `log_scale`
in eager mode. It is less error prone to make entropy models stateless and
re-instantiate them in the loop, like `tfp.distributions` objects. After
training, the user would make another instance using `compression=True` and then
share this model with sender and receiver.

In addition, this renames the generic `distribution` attribute of the entropy
model to `prior` (this doesn't necessarily indicate a Bayesian prior, but simply
a distribution that is known `a priori` to both sender and receiver).

PiperOrigin-RevId: 296313322
Change-Id: I464063b28dd1e3c0c7fa30324361ae604084bcce