Changes relative imports in models/ to package imports.
Relative imports don't work out of Colab notebooks because notebooks are
interactive and aren't run as a module. To work around this, we now prefix model
imports with the model name. This requires the model files to live in a
directory either in the Python path or in the current directory. Note that there
is a small chance that a model name clashes with a global module name, but this
risk seems unavoidable under the circumstances.

PiperOrigin-RevId: 350441806
Change-Id: I4a13fa8337b11b132142f23f89a7824db14ec88f