Updates GDN/SignalConv to TF2 API and refactors reparameterizers.
From TF 2.5, we can rely on tf.Variables contained in Layer attributes (e.g.
tf.Modules) to show up recursively in self.weights/self.trainable_weights. We
can use this here to simplify reparameterization.

With this change, we don't rely on Layer.add_weight() any more. Instead, we just
assign a tf.Variable as an attribute of the Layer. Alternatively, we allow the
attribute to be a callable (e.g. for conditioning kernels on some other
computation), or a special callable 'Parameter' which implements
reparameterization. Parameter derives from tf.Module. As a result, the contained
reparameterized tf.Variable is associated with the Layer, and will be trained.

PiperOrigin-RevId: 355009145
Change-Id: I9936142d0ce999b3ebc8f0e8e50ef6cc0a0b3e20